Refactor project-related API endpoints into modular subpackages
- Move all project-related endpoints into a nested 'projects' directory
  under 'any_auth/api/v1/', organizing by feature (members, api_keys,
  invites, and their role assignments)
- Add necessary __init__.py files to support the new package structure
- Update subrouters to use local imports and include subrouters for
  role assignments within members and api_keys
- Consolidate router inclusion in 'routes.py', removing obsolete
  imports and includes for old project-related routers
- No changes to endpoint logic; this refactor improves code
  organization, modularity, and maintainability
